==> notebooks/0_init/0_download_maxim.ipynb <==
!mkdir ../../raw
!mkdir ../../data

Files from https://www.ncbi.nlm.nih.gov/geo/download/?acc=GSE183984:

(using only .soft and two expression files supplied by publisher)

!wget https://ftp.ncbi.nlm.nih.gov/geo/series/GSE183nnn/GSE183984/soft/GSE183984_family.soft.gz
# !wget https://ftp.ncbi.nlm.nih.gov/geo/series/GSE183nnn/GSE183984/miniml/GSE183984_family.xml.tgz
# !wget https://ftp.ncbi.nlm.nih.gov/geo/series/GSE183nnn/GSE183984/matrix/GSE183984_series_matrix.txt.gz
!wget https://ftp.ncbi.nlm.nih.gov/geo/series/GSE183nnn/GSE183984/suppl/GSE183984_ASAN_RNASEQ_FPKM_ensg.csv.gz
!wget https://ftp.ncbi.nlm.nih.gov/geo/series/GSE183nnn/GSE183984/suppl/GSE183984_ASAN_RNASEQ_raw_counts_ensg.csv.gz
# !wget https://www.ncbi.nlm.nih.gov/geo/download/?type=rnaseq_counts&acc=GSE183984&format=file&file=GSE183984_raw_counts_GRCh38.p13_NCBI.tsv.gz
# !wget https://www.ncbi.nlm.nih.gov/geo/download/?type=rnaseq_counts&acc=GSE183984&format=file&file=GSE183984_norm_counts_FPKM_GRCh38.p13_NCBI.tsv.gz
# !wget https://www.ncbi.nlm.nih.gov/geo/download/?type=rnaseq_counts&acc=GSE183984&format=file&file=GSE183984_norm_counts_TPM_GRCh38.p13_NCBI.tsv.gz
# !wget https://www.ncbi.nlm.nih.gov/geo/download/?format=file&type=rnaseq_counts&file=Human.GRCh38.p13.annot.tsv.gz

!gunzip *.gz
!mv *.csv ../../raw
!mv *.soft ../../raw
# !tar -xzf GSE183984_family.xml.tgz -C ../raw && rm GSE183984_family.xml.tgz 
# !mv *.xml ../raw
# !mv *.txt ../raw
# !mv *.tsv ../raw

Genome information (maybe should use another file, for now this):

!wget ftp://ftp.ensembl.org/pub/release-113/gtf/homo_sapiens/Homo_sapiens.GRCh38.113.gtf.gz
!gzip -d Homo_sapiens.GRCh38.113.gtf.gz
!mv Homo_sapiens.GRCh38.113.gtf ../../data/Homo_sapiens.GRCh38.113.gtf
