==> notebooks/0_init/1_init_ann.ipynb <==
import os

while os.getcwd()[-len('mcrc-cetuximab-analysis'):] != 'mcrc-cetuximab-analysis':
    os.chdir('..')
os.getcwd() 

import os
import GEOparse as Geo

import pandas as pd

def decouple(s):
    ls = list(s.split("', '"))
    if len(ls) == 3:
        ls[0] = ls[0][2:]
        ls[2] = ls[2][:-2]
    elif len(ls) == 2:
        ls[0] = ls[0][2:]
        ls[1] = ls[1][:-2]
    dc = {s.split(': ')[0]: s.split(': ')[1] for s in ls}
    return dc

def extend_id(id):
    fp, sp = id.split('_')
    while len(sp) < 4:
        sp = '0' + sp
    return fp + '_' + sp

def pad_zeros(s, n):
    while len(s) < n:
        s = '0' + s
    return s

gse = Geo.get_GEO(filepath="raw/GSE183984_family.soft", silent=True)
print('Platform: ', gse.gpls)
for key in gse.gsms:
    ann = pd.DataFrame(columns=gse.gsms[key].metadata.keys())
    break

for key in gse.gsms:
    ann = pd.concat([ann, pd.DataFrame([gse.gsms[key].metadata], index=[key], dtype=str)])

print('Annotation size after collecting it from .SOFT: ', len(ann))

new_ann = pd.DataFrame(columns=['sample_id', 'tissue', 'time point', 'treatment response', 'patient_id', 'gsm_id', 'sample_title']).set_index('gsm_id')
for i in range(len(ann)):
    new_row = decouple(str(ann.iloc[i]['characteristics_ch1']))
    # print(new_row)
    title = str(ann.iloc[i]['title'])
    new_row['sample_id'] = extend_id(title[title.find('[', 1) + 1 : title.find(']')])
    new_row['sample_title'] = title[2:title.find(' ')]
    new_row['patient_id'] = (lambda s: pad_zeros(''.join(filter(lambda x: x.isdigit(), s)), 4))(new_row['sample_title'])
    new_ann = pd.concat([new_ann, pd.DataFrame([new_row], index=[ann.index[i]])])

ann = new_ann
ann['gsm_id'] = ann.index
ann = ann.set_index('sample_id')

print('Annotation size after creating new columns: ', len(ann))

ann = ann.loc[~ann["tissue"].str.contains("metastases", case=False, na=False)]
print('Annotation size after deleting metastatic samples: ', len(ann))


print('Patient 5 indexes: ', ann.loc[ann['patient_id'] == '0005'].index)
ann = ann.loc[ann['patient_id'] != '0005']
print('Annotation size after deleting patient 0005: ', len(ann))


print('Nans: ', ann.isna().sum())

# setting NaN to pre because it is known from another column (treatment response)
ann.loc[ann['time point'].isna(), 'time point'] = 'pre-treatment'

def convert_treatment_response(s):
    if 'non-PD' in s:
        return 'non-PD'
    elif 'PD' in s:
        return 'PD'
    elif s == 'pre-Tx':
        return 'unknown'
    else:
        raise Exception('unexpected treatment response in annotation row')

ann['treatment response'] = ann['treatment response'].apply(convert_treatment_response) 

treatment_response_for_patient = dict()

for i in range(len(ann)):
    row = ann.iloc[i]
    patient_id = row['patient_id']
    response = row['treatment response']
    time_point = row['time point']
    
    if time_point == 'pre-treatment' and response != 'unknown':
        raise Exception('known response in pre-treatment')
    
    if time_point == 'post-treatment' and response == 'unknown':
        raise Exception('unknown response in post-treatment')
    
    if response != 'unknown':
        if patient_id in treatment_response_for_patient and response != treatment_response_for_patient[patient_id]:
            raise Exception('contradictive post-treatment response for patient')
        
        treatment_response_for_patient[patient_id] = response

for i in range(len(ann)):
    row = ann.iloc[i]
    patient_id = row['patient_id']
    response = row['treatment response']
    
    if response == 'unknown' and patient_id in treatment_response_for_patient:
        ann.loc[ann.index[i],'treatment response'] = treatment_response_for_patient[patient_id]



ann.to_csv('data/ann_unfiltered.csv', index=True)


ann = ann.drop(['18R294_0001'])
# this is from patient 0005
# ann = ann.drop(['18R697_0005',])
print('Annotation size after deleting 18R697_0005, 18R294_0001: ', len(ann))


ann.to_csv('data/ann.csv', index=True)

ann.index

ann['tissue'].value_counts()

ann['time point'].value_counts()

ann['treatment response'].value_counts()

ann['patient_id'].value_counts().sort_index()

ann[ann['time point'] == 'post-treatment']['treatment response'].value_counts()

ann[ann['time point'] == 'pre-treatment']['treatment response'].value_counts()
